requirements.txt
numpy
Pillow
requests
simplejson
sklearn
tqdm
pyyaml
